Refactor functions into useFeedbackForm hook
