+++ client/packages/common/src/intl/locales/en/common.json
@@ -424,6 +424,7 @@
   "heading.configuration": "Configuration",
   "heading.confirm-finalise": "Are you sure you want to finalise?",
   "heading.consumption-history": "Consumption History (monthly)",
+  "heading.contact-us": "Contact Us",
   "heading.create-new-asset": "Create new asset",
   "heading.create-outbound-shipment": "Create Outbound Shipment",
   "heading.create-vaccine-course": "Create vaccine course",
@@ -1382,6 +1383,7 @@
   "messages.log-saved-successfully": "Log saved successfully",
   "messages.logout-confirm": "This will log you out.",
   "messages.max-or-min-temperature": "Max / min temperature",
+  "messages.message-not-sent": "Error! Your message was not sent",
   "messages.must-allocate-all-lines": "Cannot change the status until all lines have been allocated.",
   "messages.native-mode": "Select the mode which you would like to run. Note that you can change this option later if you have administrator access.",
   "messages.native-mode-client": "This mode allows you to select which server to connect to.",
