+++ client/packages/host/src/Help/FeedbackSection.tsx
@@ -1,8 +1,14 @@
-import { DetailContainer, Grid } from '@openmsupply-client/common';
+import {
+  DetailContainer,
+  Grid,
+  Typography,
+  useTranslation,
+} from '@openmsupply-client/common';
 import React from 'react';
 import { FeedbackForm } from './FeedbackForm';
 
 export const FeedbackSection = () => {
+  const t = useTranslation();
   return (
     <DetailContainer>
       <Grid
@@ -17,6 +23,9 @@ export const FeedbackSection = () => {
         maxWidth={800}
         gap={2}
       >
+        <Typography variant="h5" style={{ paddingBottom: 10 }}>
+          {t('heading.contact-us')}
+        </Typography>
         <FeedbackForm />
       </Grid>
     </DetailContainer>
